Updates NativeLabelScanStore on moved classes
